Refactor nunif/utils/video (tmp)
