+++ nunif/utils/video/av_filter_graph.py
@@ -1,0 +1,61 @@
+import av
+import re
+
+
+class AVFilterGraph():
+    def __init__(self, video_stream, vf, deny_filters=[]):
+        self.graph = av.filter.Graph()
+        video_filters = self.parse_vf_option(vf)
+        video_filters = [(name, option) for name, option in video_filters if name not in deny_filters]
+        self.build_graph(self.graph, video_stream, video_filters)
+
+    def update(self, frame):
+        self.graph.push(frame)
+        try:
+            return self.graph.pull()
+        except av.error.BlockingIOError:
+            return None
+        except av.error.EOFError:
+            # finished
+            return None
+
+    def flush(self):
+        out_frames = []
+        while True:
+            frame = self.update(None)
+            if frame is not None:
+                out_frames.append(frame)
+            else:
+                break
+        return out_frames
+
+    @staticmethod
+    def parse_vf_option(vf):
+        video_filters = []
+        vf = vf.strip()
+        if not vf:
+            return video_filters
+
+        for line in re.split(r'(?<!\\),', vf):
+            line = line.strip()
+            if line:
+                col = re.split(r'(?<!\\)=', line, 1)
+                if len(col) == 2:
+                    filter_name, filter_option = col
+                else:
+                    filter_name, filter_option = col[0], ""
+                filter_name, filter_option = filter_name.strip(), filter_option.strip()
+                video_filters.append((filter_name, filter_option))
+        return video_filters
+
+    @staticmethod
+    def build_graph(graph, template_stream, video_filters):
+        buffer = graph.add_buffer(template=template_stream)
+        prev_filter = buffer
+        for filter_name, filter_option in video_filters:
+            new_filter = graph.add(filter_name, filter_option if filter_option else None)
+            prev_filter.link_to(new_filter)
+            prev_filter = new_filter
+        buffersink = graph.add("buffersink")
+        prev_filter.link_to(buffersink)
+        graph.configure()
